Theta. Fixed CW dashboard after refactoring
